+++ packages/user-onboarding/projects/user-onboarding-lib/src/models/tour-cancel.ts
@@ -1,0 +1,7 @@
+import Shepherd from 'shepherd.js';
+
+export type TourCancel = {
+  index: number;
+  tour: Shepherd.Tour;
+  tourId: string;
+};