+++ services/feature-toggle-service/src/__tests__/fixtures/bearer-token-verifier.provider.ts
@@ -1,0 +1,16 @@
+import {Provider} from '@loopback/context';
+import {verify} from 'jsonwebtoken';
+import {VerifyFunction} from 'loopback4-authentication';
+import {IAuthUserWithPermissions} from '@sourceloop/core';
+
+export class BearerTokenVerifyProvider
+  implements Provider<VerifyFunction.BearerFn>
+{
+  value(): VerifyFunction.BearerFn {
+    return async (token: string) => {
+      return verify(token, 'kdskssdkdfs', {
+        issuer: 'sf',
+      }) as IAuthUserWithPermissions;
+    };
+  }
+}