+++ services/feature-toggle-service/src/__tests__/unit/user-strategy.provider.unit.ts
@@ -1,0 +1,70 @@
+import {UserStrategyProvider} from '../../providers';
+import {expect, createStubInstance} from '@loopback/testlab';
+import {StrategyHelperService} from '../../services';
+import sinon from 'sinon';
+import {StrategyBindings} from '../../keys';
+import {IAuthUserWithPermissions} from 'loopback4-authorization';
+import {
+  FeatureRepository,
+  FeatureToggleRepository,
+  StrategyRepository,
+} from '../../repositories';
+import {Getter} from '@loopback/core';
+import {FeatureFlagMetadata} from '../../types';
+import {ILogger} from '@sourceloop/core';
+import {LogMessage} from '@sourceloop/core/dist/components/logger-extension/types';
+
+describe('User Strategy provider', () => {
+  let userStrategyProvider: UserStrategyProvider;
+  let strategyHelperService: StrategyHelperService;
+  let metadataGetterStub: Getter<FeatureFlagMetadata>;
+  let user: IAuthUserWithPermissions;
+  let featureRepo: FeatureRepository;
+  let strategyRepo: StrategyRepository;
+  let featureToggleRepo: FeatureToggleRepository;
+  let logger: ILogger;
+
+  afterEach(() => sinon.restore());
+  beforeEach(setUp);
+
+  it('return true if feature is enabled at User level', async () => {
+    const isEnabled = userStrategyProvider.value();
+    const result = await isEnabled();
+    expect(result).to.be.eql(true);
+  });
+
+  function setUp() {
+    metadataGetterStub = sinon.stub().resolves({
+      featureKey: 'test',
+      strategies: [StrategyBindings.USER_STRATEGY],
+    });
+    user = {
+      permissions: ['*'],
+      authClientId: 123,
+      role: 'test_role',
+      firstName: 'test_first_name',
+      lastName: 'test_last_name',
+      username: 'test_username',
+    };
+    featureRepo = createStubInstance(FeatureRepository);
+    strategyRepo = createStubInstance(StrategyRepository);
+    featureToggleRepo = createStubInstance(FeatureToggleRepository);
+    logger = {
+      log(info: LogMessage): void {},
+      info(msg: string, key?: string): void {},
+      warn(msg: string, key?: string): void {},
+      error(msg: string, key?: string): void {},
+      debug(msg: string, key?: string): void {},
+    };
+
+    strategyHelperService = new StrategyHelperService(
+      metadataGetterStub,
+      user,
+      featureRepo,
+      featureToggleRepo,
+      strategyRepo,
+      logger,
+    );
+    userStrategyProvider = new UserStrategyProvider(strategyHelperService);
+  }
+});