+++ services/feature-toggle-service/src/decorators/feature-flag.decorator.ts
@@ -6,7 +6,8 @@ export function featuresFlag(metadata: FeatureFlagMetadata) {
   return MethodDecoratorFactory.createDecorator<FeatureFlagMetadata>(
     FEATURE_FLAG_METADATA_ACCESSOR,
     {
-      features: metadata.features || [],
+      featureKey: metadata.featureKey,
+      strategies: metadata.strategies || [],
       options: metadata.options ?? {},
     },
   );