+++ services/feature-toggle-service/src/repositories/index.ts
@@ -1,3 +1,3 @@
-export * from './features.repository';
-export * from './projects.repository';
-export * from './strategies.repository';
+export * from './feature.repository';
+export * from './strategy.repository';
+export * from './feature-toggle.repository';