+++ sandbox/payment-example/backend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.1.4 (2022-03-30)</small>
+
+* chore(deps): bump node-forge in /sandbox/payment-example/backend (#620) ([e3052fe](https://github-personal/sourcefuse/loopback4-microservice-catalog/commit/e3052fe)), closes [#620](https://github-personal/sourcefuse/loopback4-microservice-catalog/issues/620)
+
+
+
+
+
 ## <small>0.1.3 (2022-03-13)</small>
 
 **Note:** Version bump only for package @sourceloop/payment-example-backend