+++ sandbox/video-conferencing-ms-example/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>1.0.9 (2022-03-30)</small>
+
+* chore(deps): bump swagger-ui-dist (#611) ([e062e48](https://github-personal/sourcefuse/loopback4-microservice-catalog/commit/e062e48)), closes [#611](https://github-personal/sourcefuse/loopback4-microservice-catalog/issues/611)
+
+
+
+
+
 ## <small>1.0.8 (2022-03-13)</small>
 
 **Note:** Version bump only for package @sourceloop/video-conferencing-ms-example