+++ services/payment-service/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>3.2.4 (2022-03-30)</small>
+
+**Note:** Version bump only for package @sourceloop/payment-service
+
+
+
+
+
 ## <small>3.2.3 (2022-03-13)</small>
 
 * chore(deps): bump urijs in /services/payment-service (#590) ([8b4ad82](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/8b4ad82)), closes [#590](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/590)