+++ services/scheduler-service/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>4.0.7 (2022-03-30)</small>
+
+**Note:** Version bump only for package @sourceloop/scheduler-service
+
+
+
+
+
 ## <small>4.0.6 (2022-03-13)</small>
 
 * chore(deps): bump urijs in /services/scheduler-service (#584) ([d4c609b](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/d4c609b)), closes [#584](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/584)